# swagger-spec.yml
info:
  title: "Node Express Server"
  version: "1.0.0"
  description: "Node Express Server With User Creation, Authentication, Password Reset using Email Verification and JWT's."
openapi: "3.0.0"
paths:
  "/user/users":
    get:
      description: "Gets all users."
      tags:
        - User
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
              example:
                success: true
                message: "Successfully fetched all users."
                data: []
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
              example:
                success: false
                message: "An error occurred. Error: something went wrong."
                data: null
  "/user/authenticate":
    post:
      description: "Authenticate with a JWT"
      tags:
        - User
      requestBody:
        description: "A JWT to authenticate with"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
            example:
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InVzZXIiLCJpYXQiOjE1OTYwNzQwNzcsImV4cCI6MTU5NjA3NzY3N30.5f3b3b3b3b3b3b3b3b3b3b3b3b"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  userExists:
                    type: boolean
                    example: true
                  loggedIn:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "User is logged in."
                  data:
                    description: "The users MongoDB document _id or undefined if no doc."
                    type: string
                    nullable: true
                    example: "5f3b3b3b3b3b3b3b3b3b3b3b"

        "422":
          description: "Invalid request body"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  userExists:
                    type: boolean
                    example: false
                  loggedIn:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Invalid request body."
                  data:
                    description: "The users MongoDB document _id or undefined if no doc."
                    type: string
                    nullable: true
                    example: null
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  userExists:
                    type: boolean
                    example: false
                  loggedIn:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "An error occurred. Error: something went wrong."
                  data:
                    description: "The users MongoDB document _id or undefined if no doc."
                    type: string
                    nullable: true
                    example: null
...
  "/user/email-available":
    post:
      description: "Checks if an email is available."
      tags:
        - User
      requestBody:
        description: "The email to query with"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
            example:
              email: "my@email.com"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  emailAvailable:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Email is available."
        "422":
          description: "Invalid request body"
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  emailAvailable:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Invalid request body."
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  emailAvailable:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "An error occurred. Error: something went wrong."
  "/user/forgot-password":
    post:
      description: Sends a password reset email.
      tags:
        - User
      requestBody:
        description: the email address of the account to request a password reset email to.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
            example:
              email: my@email.com
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Password reset email sent."
                  data:
                    description: "The users MongoDB document _id or undefined if no doc."
                    type: string
                    nullable: true
                    example: "5f3b3b3b3b3b3b3b3b3b3b3b"
        "422":
          description: Invalid request body
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Invalid request body."
                  data:
                    description: "The users MongoDB document _id or undefined if no doc."
                    type: string
                    nullable: true
                    example: null
        "500":
          description: Caught error
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "An error occurred. Error: something went wrong."
                  data:
                    description: "The users MongoDB document _id or undefined if no doc."
                    type: string
                    nullable: true
                    example: null
...
components:
  schemas:
...
    User:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        journals:
          type: array
          items:
            $ref: "#/components/schemas/Journal"
        journalCategories:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              selected:
                type: boolean
    Journal:
      type: object
      properties:
        title:
          type: string
          example: "Journal 1"
        entry:
          type: string
          example: "Journal 1"
        category:
          type: string
        date:
          type: string
        selected:
          type: boolean
...